+++ Changelog.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.2.32] - 2024-04-13
+- include more instructions in `--simplify`
+- handle combinations like `--mca --intel` in addition to `--mca-intel` which is now deprecated
+- cosmetic improvements in produced output
+- a bunch of internal improvements
+- drop `once_cell` dependency
+
 ## [0.2.31] - 2024-04-04
 - include relevant constants in produced asm output
   this can be disabled with `--no-constants`